feat: create MVC file structure & add .env file
