+++ .gitignore
@@ -1,2 +1,3 @@
+.env
 .idea/
 __pycache__/
