+++ src/controller/recommendations.py
@@ -1,0 +1,7 @@
+@app.get("/recommendations/user")
+def get_user_recommendations():
+    return {"message": "User recommendations"}
+
+@app.get("/recommendations/movie/{movie_id}")
+def get_movie_recommendations(movie_id: int):
+    return {"message": f"Recommendations for movie {movie_id}"}
