+++ src/model/tables.py
@@ -1,0 +1,38 @@
+from sqlalchemy import create_engine, Column, Integer, String, Float, ForeignKey
+from sqlalchemy.ext.declarative import declarative_base
+from sqlalchemy.orm import sessionmaker, Session, relationship
+
+
+# database setup
+DATABASE_URL = "postgresql://username:password@localhost/recommender"
+engine = create_engine(DATABASE_URL)
+SessionLocal = sessionmaker(autocommit=False, autoflush=False, bind=engine)
+Base = declarative_base()
+
+class User(Base):
+    __tablename__ = "users"
+    id = Column(Integer, primary_key=True, index=True)
+    username = Column(String, unique=True, index=True)
+    password_hash = Column(String)
+
+class Movie(Base):
+    __tablename__ = "movies"
+    id = Column(Integer, primary_key=True, index=True)
+    title = Column(String, index=True)
+    genre = Column(String)
+    overview = Column(String)
+    rating = Column(Float)
+    director = Column(String)
+    actors = Column(String)
+
+class Rating(Base):
+    __tablename__ = "ratings"
+    id = Column(Integer, primary_key=True, index=True)
+    user_id = Column(Integer, ForeignKey("users.id"))
+    movie_id = Column(Integer, ForeignKey("movies.id"))
+    rating = Column(Float)
+
+    user = relationship("User")
+    movie = relationship("Movie")
+
+Base.metadata.create_all(bind=engine)
